requirements.txt
PySide6>=6.6.0
google-genai>=1.0.0
pydub>=0.25.1
# Note: pydub requires ffmpeg to be installed and available in PATH for MP3 export.
# Download ffmpeg from https://ffmpeg.org/download.html and add to your system PATH.
